Changing two rules
The first one is related to the use of "var", which was used once for multiple variables with short assignments and multiple times for variables with long assignments. For assimilation sake, established the use of var always as the correct way to do it.

The second one, related to IF, is just to avoid changing the if statement (i.e. breaking line and adding brackets) when needing to add more code to it.